RavenDB-13113: Move namespaces and rebase
